refactor: move marquee list and speed animation into framework/ui
