+++ src/components/list/index.js
@@ -3,5 +3,4 @@ export { default as ListFooter } from './ListFooter';
 export { default as ListHeader, ListHeaderHeight } from './ListHeader';
 export { default as ListItem } from './ListItem';
 export { default as ListItemDivider } from './ListItemDivider';
-export { default as MarqueeList } from './MarqueeList';
 export { default as NoResults } from './NoResults';
